+++ src/components/common/CustomModal.ts
@@ -1,4 +1,4 @@
-import { $, dispatchCustomEvent } from "../utils/dom";
+import { $, dispatchCustomEvent } from "../../utils/dom";
 
 class CustomModal extends HTMLElement {
   constructor() {
