+++ src/components/common/ObservedArea.ts
@@ -1,4 +1,4 @@
-import { dispatchCustomEvent, $ } from "../utils/dom";
+import { dispatchCustomEvent, $ } from "../../utils/dom";
 
 class ObservedArea extends HTMLElement {
   constructor() {
