+++ src/components/header/SearchBox.ts
@@ -1,4 +1,4 @@
-import { $, dispatchCustomEvent } from "./../utils/dom";
+import { $, dispatchCustomEvent } from "../../utils/dom";
 
 class SearchBox extends HTMLElement {
   constructor() {
