+++ src/components/movie/MovieListContainer.ts
@@ -1,4 +1,4 @@
-import { $ } from "./../utils/dom";
+import { $ } from "../../utils/dom";
 
 class MovieListContainer extends HTMLElement {
   constructor() {
