+++ src/index.js
@@ -1,15 +1,15 @@
 import "../templates/reset.css";
 import "../templates/common.css";
 
-import MovieHeader from "./components/MovieHeader";
-import SearchBox from "./components/SearchBox";
-import MovieItem from "./components/MovieItem";
-import MovieList from "./components/MovieList";
-import MovieListContainer from "./components/MovieListContainer";
-import CustomModal from "./components/CustomModal";
-import MovieDetail from "./components/MovieDetail";
-import MovieScore from "./components/MovieScore";
-import ObservedArea from "./components/ObservedArea";
+import MovieHeader from "./components/header/MovieHeader";
+import SearchBox from "./components/header/SearchBox";
+import MovieItem from "./components/movie/MovieItem";
+import MovieList from "./components/movie/MovieList";
+import MovieListContainer from "./components/movie/MovieListContainer";
+import CustomModal from "./components/common/CustomModal";
+import MovieDetail from "./components/movie/MovieDetail";
+import MovieScore from "./components/movie/MovieScore";
+import ObservedArea from "./components/common/ObservedArea";
 
 import movieApp from "./movieApp";
